+++ modules/apps/portal-background-task/portal-background-task-service/src/main/java/com/liferay/portal/background/task/internal/upgrade/v1_0_0/SchemaUpgradeProcess.java
@@ -14,20 +14,24 @@
 
 package com.liferay.portal.background.task.internal.upgrade.v1_0_0;
 
-import com.liferay.portal.background.task.internal.upgrade.v1_0_0.util.BackgroundTaskTable;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
+import com.liferay.portal.kernel.util.LoggingTimer;
+import com.liferay.portal.kernel.util.StringUtil;
 
 /**
- * @author Cristina González
+ * @author Miguel Pastor
  */
-public class UpgradeBackgroundTask extends UpgradeProcess {
+public class SchemaUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		alter(
-			BackgroundTaskTable.class,
-			new AlterColumnType("name", "VARCHAR(255) null"),
-			new AlterColumnName("taskContext", "taskContextMap TEXT null"));
+		try (LoggingTimer loggingTimer = new LoggingTimer()) {
+			String template = StringUtil.read(
+				SchemaUpgradeProcess.class.getResourceAsStream(
+					"dependencies/update.sql"));
+
+			runSQLTemplateString(template, false);
+		}
 	}
 
 }
