+++ modules/apps/portal-scheduler/portal-scheduler-quartz/src/main/java/com/liferay/portal/scheduler/quartz/internal/upgrade/QuartzServiceUpgrade.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.scheduler.quartz.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
-import com.liferay.portal.scheduler.quartz.internal.upgrade.v1_0_0.UpgradeQuartz;
+import com.liferay.portal.scheduler.quartz.internal.upgrade.v1_0_0.QuartzUpgradeProcess;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
@@ -32,7 +32,7 @@ public class QuartzServiceUpgrade implements UpgradeStepRegistrator {
 	public void register(Registry registry) {
 		registry.register("0.0.0", "1.0.0", new DummyUpgradeStep());
 
-		registry.register("0.0.1", "1.0.0", new UpgradeQuartz());
+		registry.register("0.0.1", "1.0.0", new QuartzUpgradeProcess());
 	}
 
 }
