+++ modules/apps/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/upgrade/SearchWebUpgrade.java
@@ -17,7 +17,7 @@ package com.liferay.portal.search.web.internal.upgrade;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.search.web.internal.upgrade.v1_0_0.UpgradePortletId;
 import com.liferay.portal.search.web.internal.upgrade.v1_0_0.UpgradePortletPreferences;
-import com.liferay.portal.search.web.internal.upgrade.v2_0_0.UpgradeSearchPortlet;
+import com.liferay.portal.search.web.internal.upgrade.v2_0_0.SearchPortletUpgradeProcess;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
@@ -36,7 +36,7 @@ public class SearchWebUpgrade implements UpgradeStepRegistrator {
 			"0.0.1", "1.0.0", new UpgradePortletId(),
 			new UpgradePortletPreferences());
 
-		registry.register("1.0.0", "2.0.0", new UpgradeSearchPortlet());
+		registry.register("1.0.0", "2.0.0", new SearchPortletUpgradeProcess());
 	}
 
 }
