+++ modules/apps/portal-security-sso/portal-security-sso-token-impl/src/main/java/com/liferay/portal/security/sso/token/internal/upgrade/v2_0_0/TokenConfigurationUpgradeProcess.java
@@ -35,7 +35,7 @@ import java.util.List;
 /**
  * @author Christopher Kian
  */
-public class UpgradeTokenConfiguration extends UpgradeProcess {
+public class TokenConfigurationUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
