+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/upgrade/PortalSettingWebUpgrade.java
@@ -17,7 +17,7 @@ package com.liferay.portal.settings.web.internal.upgrade;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.settings.web.internal.upgrade.v1_0_0.UpgradePortletId;
 import com.liferay.portal.settings.web.internal.upgrade.v1_0_1.UpgradeInstanceSettingsPortletId;
-import com.liferay.portal.settings.web.internal.upgrade.v1_0_2.UpgradeResourceAction;
+import com.liferay.portal.settings.web.internal.upgrade.v1_0_2.ResourceActionUpgradeProcess;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
@@ -37,7 +37,7 @@ public class PortalSettingWebUpgrade implements UpgradeStepRegistrator {
 		registry.register(
 			"1.0.0", "1.0.1", new UpgradeInstanceSettingsPortletId());
 
-		registry.register("1.0.1", "1.0.2", new UpgradeResourceAction());
+		registry.register("1.0.1", "1.0.2", new ResourceActionUpgradeProcess());
 	}
 
 }
