+++ modules/apps/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_1_0/WorkflowContextUpgradeProcess.java
@@ -41,7 +41,7 @@ import org.json.JSONObject;
 /**
  * @author Jang Kim
  */
-public class UpgradeWorkflowContext extends UpgradeProcess {
+public class WorkflowContextUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
