+++ modules/apps/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_3_0/KaleoDefinitionUpgradeProcess.java
@@ -23,7 +23,7 @@ import java.sql.ResultSet;
 /**
  * @author Leonardo Barros
  */
-public class UpgradeKaleoDefinition extends UpgradeProcess {
+public class KaleoDefinitionUpgradeProcess extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
