+++ modules/apps/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_3_2/KaleoClassNameAndKaleoClassPKUpgradeProcess.java
@@ -24,7 +24,8 @@ import java.sql.PreparedStatement;
 /**
  * @author Rafael Praxedes
  */
-public class UpgradeKaleoClassNameAndKaleoClassPK extends UpgradeProcess {
+public class KaleoClassNameAndKaleoClassPKUpgradeProcess
+	extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
