+++ modules/apps/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_4_1/KaleoDefinitionVersionUpgradeProcess.java
@@ -36,7 +36,7 @@ import java.util.List;
 /**
  * @author Inácio Nery
  */
-public class UpgradeKaleoDefinitionVersion extends UpgradeProcess {
+public class KaleoDefinitionVersionUpgradeProcess extends UpgradeProcess {
 
 	protected void addBatch(
 			PreparedStatement ps, long kaleoDefinitionId,
