+++ modules/apps/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v3_0_0/SchemaUpgradeProcess.java
@@ -48,7 +48,7 @@ import java.util.List;
 /**
  * @author Inácio Nery
  */
-public class UpgradeSchema extends UpgradeProcess {
+public class SchemaUpgradeProcess extends UpgradeProcess {
 
 	protected void addBatch(
 			PreparedStatement ps, long kaleoDefinitionId,
