+++ modules/apps/portal/portal-upload/src/main/java/com/liferay/portal/upload/internal/upgrade/UploadUpgrade.java
@@ -17,7 +17,7 @@ package com.liferay.portal.upload.internal.upgrade;
 import com.liferay.portal.configuration.upgrade.PrefsPropsToConfigurationUpgradeHelper;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.upload.internal.upgrade.v1_0_0.UpgradeUploadServletRequestConfiguration;
+import com.liferay.portal.upload.internal.upgrade.v1_0_0.UploadServletRequestConfigurationUpgradeProcess;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -34,7 +34,7 @@ public class UploadUpgrade implements UpgradeStepRegistrator {
 
 		registry.register(
 			"0.0.1", "1.0.0",
-			new UpgradeUploadServletRequestConfiguration(
+			new UploadServletRequestConfigurationUpgradeProcess(
 				_prefsPropsToConfigurationUpgradeHelper));
 	}
