+++ modules/apps/product-navigation/product-navigation-product-menu-web/src/main/java/com/liferay/product/navigation/product/menu/web/internal/upgrade/ProductNavigationProductMenuWebUpgrade.java
@@ -16,7 +16,7 @@ package com.liferay.product.navigation.product.menu.web.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.product.navigation.product.menu.web.internal.upgrade.v1_0_0.UpgradePortletPreferences;
+import com.liferay.product.navigation.product.menu.web.internal.upgrade.v1_0_0.PortletPreferencesUpgradeProcess;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -31,7 +31,8 @@ public class ProductNavigationProductMenuWebUpgrade
 	public void register(Registry registry) {
 		registry.register("0.0.0", "1.0.0", new DummyUpgradeStep());
 
-		registry.register("0.0.1", "1.0.0", new UpgradePortletPreferences());
+		registry.register(
+			"0.0.1", "1.0.0", new PortletPreferencesUpgradeProcess());
 	}
 
 }
