+++ modules/apps/product-navigation/product-navigation-product-menu-web/src/main/java/com/liferay/product/navigation/product/menu/web/internal/upgrade/v1_0_0/PortletPreferencesUpgradeProcess.java
@@ -12,24 +12,34 @@
  * details.
  */
 
-package com.liferay.portal.security.audit.storage.internal.upgrade.v1_0_1;
+package com.liferay.product.navigation.product.menu.web.internal.upgrade.v1_0_0;
 
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.LoggingTimer;
-import com.liferay.portal.security.audit.storage.internal.upgrade.v1_0_1.util.AuditEventTable;
 
 /**
- * @author Peter Petrekanics
+ * @author Jürgen Kappler
  */
-public class UpgradeSchema extends UpgradeProcess {
+public class PortletPreferencesUpgradeProcess extends UpgradeProcess {
+
+	protected void deletePortletPreferences(String portletId) throws Exception {
+		if (_log.isDebugEnabled()) {
+			_log.debug("Delete portlet preferences for portlet " + portletId);
+		}
+
+		runSQL(
+			"delete from PortletPreferences where portletId = '" + portletId +
+				"'");
+	}
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		try (LoggingTimer loggingTimer = new LoggingTimer()) {
-			alter(
-				AuditEventTable.class,
-				new AlterColumnType("clientIP", "VARCHAR(255) null"));
-		}
+		deletePortletPreferences("145");
+		deletePortletPreferences("160");
 	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		PortletPreferencesUpgradeProcess.class);
 
 }
