+++ modules/apps/redirect/redirect-service/src/main/java/com/liferay/redirect/internal/upgrade/RedirectServiceUpgrade.java
@@ -29,17 +29,17 @@ public class RedirectServiceUpgrade implements UpgradeStepRegistrator {
 		registry.register(
 			"1.0.0", "2.0.0",
 			new com.liferay.redirect.internal.upgrade.v2_0_0.
-				UpgradeRedirectNotFoundEntry());
+				RedirectNotFoundEntryUpgradeProcess());
 
 		registry.register(
 			"2.0.0", "2.0.1",
 			new com.liferay.redirect.internal.upgrade.v2_0_1.
-				UpgradeRedirectNotFoundEntry());
+				RedirectNotFoundEntryUpgradeProcess());
 
 		registry.register(
 			"2.0.1", "3.0.0",
 			new com.liferay.redirect.internal.upgrade.v3_0_0.
-				UpgradeRedirectNotFoundEntry());
+				RedirectNotFoundEntryUpgradeProcess());
 	}
 
 }
